Update the next-config to take env variablesm rename swiper (#41)
* Update the next-config to take env variablesm rename swiper
# Conflicts:
#	src/app/custom-questions/GameRoomClient.tsx
#	src/app/custom-questions/Lobby.tsx
#	src/app/custom-questions/Room.tsx
#	src/app/custom-questions/page.tsx

* Update the next-config to take env variablesm rename swiper
# Conflicts:
#	src/app/custom-questions/GameRoomClient.tsx
#	src/app/custom-questions/Lobby.tsx
#	src/app/custom-questions/Room.tsx
#	src/app/custom-questions/page.tsx
